third_party/blink/web_tests/platform/mac/virtual/forced-high-contrast-colors/external/wpt/forced-colors-mode/forced-colors-mode-04-expected.txt

This is a testharness.js-based test.
FAIL Checks hyperlinks are overridden in forced colors mode. assert_equals: expected "rgba(0, 0, 0, 0.18)" but got "rgba(0, 0, 0, 0.4)"
Harness: the test ran to completion.
